Tidy Docker package.
